Use a different structure for protocol and user votes
Since protocol voting the vote serialization doesn't include the version
anymore. It's a problem for the user vote that is stored in the wallet
because we must be able to load old votes during upgrades.

This commit adds a new CUserVote structure that inherits CVote and adds
a serialized version.

In the future these two structures may differ more.