Keep only a part of the block indexes in memory
Use a special BlockMap class for mapBlockIndex that loads block indexes
on demand.

Two other CBlockIndex* like classes are added:

CLazyBlockIndex: Constructed from a hash it behaves like a CBlockIndex*
but only loads data when it's used. If the hash is 0 it behaves like a
NULL pointer. It's used as pprev and pnext in CBlockIndex.

CLockedBlockIndex: Behaves like a CBlockIndex but additionnally prevents
the real CBlockIndex* from being unloaded.

The block index table is cleaned after a block is received and regularly
during the few processes that need to scan the whole blockchain.

During the cleanup the most recent blocks are kept in memory.

Some CBlockIndex members that were only kept in memory have been added
to the block index database.